refactor: make app its own component
